requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.31
asyncpg==0.31.0
alembic==1.18.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic[email]==2.12.5
psycopg2-binary
python-dotenv==1.0.0